Created Separate Class Library for Backend Classes and moved Item, Drink, Recipe, Inventory, and DatabaseManager into said Class Library, and added references to it in the main BRIM project. NOTE: code compiles and works fine (at least the back end part does when I tested it running Program.Main") but classes in BackendClassLibrary will still throw errors about conflicting with themselves. When I looked online, the answers where either that it was something to ignore, or possibly the result of circular references within one of the projects. Either way, It's not something I can solve, as inuitive C# project management is not one of VScode's strengths
