+++ BRIM.csproj
@@ -24,6 +24,10 @@
     <PackageReference Include="System.Data.SqlClient" Version="4.8.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="BackendClassLibrary\BackendClassLibrary.csproj" />
+  </ItemGroup>
+
   <Target Name="ReactBuild" AfterTargets="Build">
     <Exec Command="npm i &amp;&amp; npm run build" />
   </Target>
