+++ BRIM.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BRIM", "BRIM.csproj", "{7B5E82B8-388E-4371-8442-36F61C085767}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BackendClassLibrary", "BackendClassLibrary\BackendClassLibrary.csproj", "{4663DF12-FDB1-44D1-BDD2-271F4154F596}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{7B5E82B8-388E-4371-8442-36F61C085767}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7B5E82B8-388E-4371-8442-36F61C085767}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7B5E82B8-388E-4371-8442-36F61C085767}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4663DF12-FDB1-44D1-BDD2-271F4154F596}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4663DF12-FDB1-44D1-BDD2-271F4154F596}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4663DF12-FDB1-44D1-BDD2-271F4154F596}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4663DF12-FDB1-44D1-BDD2-271F4154F596}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
