+++ BackendClassLibrary/DatabaseManager.cs
@@ -6,7 +6,7 @@ using System.Data.SqlClient;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace BRIM
+namespace BackendClassLibrary
 {
     public class DatabaseManager
     {
