+++ BackendClassLibrary/Drink.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Data;
 
-namespace BRIM
+namespace BackendClassLibrary
 {
     public class Drink : Item
     {
